refactor p4
