+++ src/main/java/com/jredfox/lanessentials/commands/client/CommandPublicIP.java
@@ -1,6 +1,4 @@
-package com.jredfox.lanessentials.client;
-
-import java.net.UnknownHostException;
+package com.jredfox.lanessentials.commands.client;
 
 import com.evilnotch.lib.minecraft.util.EnumChatFormatting;
 import com.evilnotch.lib.minecraft.util.PlayerUtil;
@@ -13,7 +11,7 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.server.MinecraftServer;
 import net.minecraftforge.client.IClientCommand;
 
-public class CommandIP extends CommandBase implements IClientCommand{
+public class CommandPublicIP extends CommandBase implements IClientCommand{
 	
     /**
      * Return the required permission level for this command.
@@ -22,15 +20,15 @@ public class CommandIP extends CommandBase implements IClientCommand{
     {
         return 0;
     }
-
+	
 	@Override
 	public String getName() {
-		return "ipv4";
+		return "publicIp";
 	}
 
 	@Override
 	public String getUsage(ICommandSender sender) {
-		return "/ipv4";
+		return "/publicIp";
 	}
 
 	@Override
@@ -40,13 +38,13 @@ public class CommandIP extends CommandBase implements IClientCommand{
 			return;
 		try 
 		{
-			String ip = JavaUtil.getIpv4();
-			PlayerUtil.sendClipBoard((EntityPlayer)sender, "Your Ipv4 Adress is:",  EnumChatFormatting.BLUE + ip);
+			String ip = JavaUtil.getPublicIp();
+			PlayerUtil.sendClipBoard((EntityPlayer)sender, "Your Public Ip Adress is:", ip);
 			PlayerUtil.copyClipBoard((EntityPlayer)sender, ip);
 		} 
-		catch (UnknownHostException e) 
+		catch(Exception e) 
 		{
-			PlayerUtil.printChat((EntityPlayer)sender, EnumChatFormatting.RED, "", "Unable to fetch ipv4 adress");
+			PlayerUtil.printChat((EntityPlayer)sender, EnumChatFormatting.RED, "", "Unable to fetch public ip adress");
 		}
 	}
 
@@ -54,6 +52,5 @@ public class CommandIP extends CommandBase implements IClientCommand{
 	public boolean allowUsageWithoutPrefix(ICommandSender sender, String message) {
 		return true;
 	}
-	
 
 }
