+++ src/main/java/com/jredfox/lanessentials/events/LEEventHandler.java
@@ -10,7 +10,6 @@ import com.jredfox.lanessentials.packets.PacketNick;
 
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
-import net.minecraft.network.NetHandlerPlayServer;
 import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
